Add KubeDeploy handler
Add ImageLinked event handler that deploys to kube.  It uses the
GoogleContainerBuilder status on the commit as an eligibility flag.
If no existing deployment exists, create service and deployment.  If
one exists, update the image.

Remove unused k8 resource specs, the templates are in code.

Ingress still to do.

Towards #1