Create ingress rule for deployment service
A single global ingress is in the default namespace, see
assets/gke/ingress.json, so it can only forward requests to services
in that namespace.  Create two services for each deployment, one in
its namespace and one in default.  Manually create endpoints for the
service in default that forward to the service in the deployment
namespace.

Use the Atomist team ID for namespacing, and use repo owner in
resource name.  Fix kubernetes resource names to be valid, i.e, remove
'.'.

Convert templates to functions to avoid the stringify-parse-modify
badness.