Add rate limiting, remove host from ingress
So we can only provide a single secret in the ingress-nginx namespace
providing the TLS cert and key, we remove the host from the ingress
rules so everything routes through the default host on the ingress,
which will use the default TLS config and redirect from HTTP to HTTPS.

Add rate limiting annotations to ingress.

Fix IngressSpec.

Towards atomist/github-sdm#94