Use default merge for Kubernetes specs
Rather than overriding the default behavior of lodash.merge for
arrays, allow it to deeply merge array elements.

Another option would be to create the deployment resource using the
default template and then patch it using the provided spec.  This
seems more risky because it depends on an agreed naming for
containers.