Add namespace mode support
Implement ability to run in "namespace mode" where k8-automation is
running in and managing deploys in a single namespace.  When running
in namespace mode, the namespace must be provided as the value of the
POD_NAMESPACE environment variable.  Make all specs consistent.  Add
k8-automation namespace to role binding when in cluster mode.  Move
goal validation into own function and add tests.